requirements.txt
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
compiler==0.2.0
cycler==0.10.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
finnhub-python==2.4.5
Flask==2.0.2
Flask-Compress==1.10.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.3
pandas==1.3.4
Pillow==8.4.0
plotly==5.3.1
pyparsing==3.0.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
tenacity==8.0.1
urllib3==1.26.7
Werkzeug==2.0.2
